Fixed bugs
GetItNow is an innovative ecommerce website that takes online shopping to a whole new level. Unlike traditional platforms, we incorporated the power of three.js library to introduce 3D models of the listed items. This feature allows users to have an immersive and interactive experience while exploring products. And guess what? We have even bigger plans to take this concept into the realm of Virtual Reality (VR) soon!